key can only be get by maincharacter
