+++ kagefumi/Assets/Scripts/Game/Character.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: d56929b8b743a4087a21d1b5fc720295
+folderAsset: yes
+timeCreated: 1449900380
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
