+++ kagefumi/Assets/Scripts/Game/Character/MainCharacterController.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 6afda0510b7a44eca81ced5bfafc309f
+timeCreated: 1449900393
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
